Rename files to fit Rubocop gem layout
