--- lib/custom/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Custom
-  class NoMagicNumbers
-    VERSION = '0.0.1'
-  end
-end
