+++ lib/rubocop/no_magic_numbers.rb
@@ -1,0 +1,1 @@
+require "rubocop/cop/magic_numbers/no_magic_numbers"
